backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-polymorphic
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
